+++ TabletopTweaks-Base/Bugfixes/Classes/Cavalier.cs
@@ -16,6 +16,7 @@ using Kingmaker.UnitLogic.Mechanics.Components;
 using Kingmaker.Utility;
 using System.Linq;
 using TabletopTweaks.Core.NewComponents;
+using TabletopTweaks.Core.NewComponents.OwlcatReplacements;
 using TabletopTweaks.Core.NewComponents.Prerequisites;
 using TabletopTweaks.Core.Utilities;
 using static TabletopTweaks.Base.Main;
@@ -81,6 +82,7 @@ namespace TabletopTweaks.Base.Bugfixes.Classes {
                 PatchCavalierMountSelection();
                 PatchMightyCharge();
                 PatchSupremeCharge();
+                PatchOrderOfTheCockatrice();
                 PatchOrderOfTheStar();
 
                 void PatchMightyCharge() {
@@ -152,6 +154,24 @@ namespace TabletopTweaks.Base.Bugfixes.Classes {
 
                     CavalierProgression.LevelEntries.Where(l => l.Level == 1).First().m_Features.Add(CavalierMobilityFeature.ToReference<BlueprintFeatureBaseReference>());
                     DiscipleOfThePikeArchetype.RemoveFeatures.Where(l => l.Level == 1).First().m_Features.Add(CavalierMobilityFeature.ToReference<BlueprintFeatureBaseReference>());
+                }
+                void PatchOrderOfTheCockatrice() {
+                    PatchChallenge();
+
+                    void PatchChallenge() {
+                        if (TTTContext.Fixes.Cavalier.Base.IsDisabled("OrderOfTheCockatriceChallenge")) { return; }
+
+                        var CavalierCockatriceChallenge = BlueprintTools.GetBlueprint<BlueprintFeature>("ba176bca404967b47bf9e583e80c3fd5");
+                        CavalierCockatriceChallenge.TemporaryContext(bp => {
+                            var oldComponent = bp.GetComponent<DamageBonusOrderOfCockatrice>();
+                            bp.AddComponent<DamageBonusOrderOfCockatriceTTT>(c => {
+                                c.m_CheckedFact = oldComponent.m_CheckedFact;
+                                c.Bonus = oldComponent.Bonus;
+                                c.Descriptor = oldComponent.Descriptor;
+                            });
+                            bp.RemoveComponents<DamageBonusOrderOfCockatrice>();
+                        });
+                    }
                 }
                 void PatchOrderOfTheStar() {
                     PatchCalling();